+++ CHANGES.rst
@@ -4,9 +4,13 @@ Changelog
 0.4 (unreleased)
 ----------------
 
-- Add session_factory_from_settings
+- Add session_factory_from_settings.
 
-- Add includeme
+- Add includeme.
+
+- Drop Python 2.6, 3.3, and 3.4 support.
+
+- Add Python 3.6, 3.7 and 3.8 support.
 
 0.3 (2016-02-16)
 ----------------