+++ CHANGES.rst
@@ -4,9 +4,8 @@ Changelog
 0.4 (unreleased)
 ----------------
 
-- Add session_factory_from_settings.
-
-- Add includeme.
+- Add ``session_factory_from_settings`` and ``includeme``.
+  See https://github.com/Pylons/pyramid_nacl_session/pull/11
 
 - Drop Python 2.6, 3.3, and 3.4 support.